pyproject.toml
[tool.poetry]
name = "bai-mkrl"
version = "0.1.0"
description = ""
authors = ["Davy Cielen <data@spencer.cash>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
transformers = "^4.36.2"
pytesseract = "^0.3.10"
datasets = "^2.16.1"
torch = "^2.1.2"
torchvision = "^0.16.2"
seqeval = "^1.2.2"
tensorboard = "^2.15.1"
jupyter = "^1.0.0"
evaluate = "^0.4.1"
accelerate = "^0.26.1"
streamlit = "^1.30.0"
langchain-core = "^0.1.12"
unstructured = {extras = ["xlsx"], version = "^0.12.0"}
python-dotenv = "^1.0.0"
langchain = "^0.1.1"
langchain-openai = "^0.0.2.post1"
langchain-community = "^0.0.13"
opencv-python = "^4.9.0.80"
unstructured-pytesseract = "^0.3.12"
pdfminer = "^20191125"
unstructured-inference = "^0.7.21"
pikepdf = "^8.11.2"
pypdf = "^3.17.4"
langchain-experimental = "^0.0.49"
idom = "^1.0.0"
reactpy = "^1.0.2"
openpyxl = "^3.1.2"
streamlit-ace = "^0.1.1"
arxiv = "^2.1.0"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core","setuptools", "wheel", "torch", "pip"]
